+++ frontend/src/pages/SurveyResponse/SurveyResponse.tsx
@@ -4,6 +4,8 @@
  import loadingIndicator from '../utils/loading.svg';
 
  import Responses from './Responses';
+ import SRHeader from './SRHeader';
+ import Properties from './Properties';
  import './images/profile_pic_default.png';
  import './SurveyResponse.css';
  
@@ -14,13 +16,7 @@ process.env.REACT_APP_backendURL || 'https://labs10-cleaner-app-2.herokuapp.com'
 
 const [response, err, loading] = useFetch(`${url}/surveyresponses/${props.match.params.id}`, true, 'get');
 
-let surveyName = () => {
-    if(response.survey.length === 0){
-        return("No Surveys Here");
-    }else{
-        return(`${response.survey[0].name} - Responses`);
-        }
-    }
+
 
 let guestPic = () => {
     let sIndex = -1;
@@ -35,28 +31,36 @@ let guestPic = () => {
     } 
 }
 
+const [matches, setMatches] = useState(false);
+let [property, setProperty] = useState("");
 
-          
+let changeHandler = (e:any) => {
+    e.preventDefault();
+    setMatches(true);
+    setProperty(property = e.target.value);
+    if(property === "title"){
+        setMatches(false);
+    }
+}
+
+// let filteredSurveys = response.survey.filter((survey:any) => {
+//     console.log(response.survey);
+//     console.log(survey);
+//     if(matches === true){
+//         return survey.house_name === property;
+//     }else{
+//         return survey;
+//     }
+// })  
+
 if (loading === true) {
     return(
         <img src = {loadingIndicator} />
-    )} else {
+    )}else{
         return(
             <div className = 'sr-container'>
-                <div className = 'sr-title'>
-                    <h1>{surveyName()}</h1>
-                    {console.log(response.survey)}
-                </div> 
-
-                <div className='sr-title'>
-                    <select name="properties">
-                        <option value="title">Filter By Property</option>
-                        {response.survey.map((survey:any) =>{
-                        return(<option value={survey.house_name}>{survey.house_name}</option> )
-                        })}
-                    </select> 
-                </div>
-                
+                <SRHeader surveys = {response.survey} />
+                <Properties surveys = {response.survey} changeHandler ={changeHandler}/> 
                 {response.survey.map((survey: any) =>
                 <Responses key={survey.id} survey={response.survey} sr_name={survey.guest_name} sr_date={survey.created_at} guestPic={guestPic()} sr_img={survey.photo} />
                 )}
